Changed flags to long to align with infinispan

+++ src/main/java/com/zalora/jmemcached/CacheElement.java
@@ -17,7 +17,7 @@ public interface CacheElement extends Serializable, SizedItem {
 
     long getExpire();
 
-    int getFlags();
+    long getFlags();
 
     ChannelBuffer getData();
